requirements.txt
beautifulsoup4==4.6.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.9
django-modelcluster==5.0.1
django-taggit==1.2.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
html5lib==1.0.1
idna==2.8
Pillow==6.2.2
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.8
wagtail==2.7.1
webencodings==0.5.1
Willow==1.3
